parser: Move Token, TokenStack, TokenType class to parser package.
